requirements.txt
dialogflow==0.8.0
absl-py==0.9.0
aenum==2.2.3
aiofiles==0.4.0
alabaster==0.7.12
APScheduler==3.6.3
asn1crypto==1.0.1
astor==0.8.1
astroid==2.3.1
astropy==3.2.1
atomicwrites==1.3.0
attrs==19.2.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bitarray==1.0.1
bkcharts==0.2
Bottleneck==1.2.1
cachetools==4.0.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.2
decorator==4.4.0
defusedxml==0.6.0
distributed==2.5.2
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
fbmessenger==6.0.0
feedfinder2==0.0.4
feedparser==5.2.1
filelock==3.0.12
Flask==1.1.1
fsspec==0.5.2
future==0.17.1
gevent==1.4.0
glob2==0.7
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.26.0
gunicorn==20.0.4
h11==0.8.1
h2==3.2.0
h5py==2.9.0
HeapDict==1.0.1
hpack==3.0.0
hstspreload==2020.3.12
html5lib==1.0.1
httptools==0.1.1
httpx==0.9.3
idna==2.8
imageio==2.6.0
imagesize==1.1.0
importlib-metadata==0.23
ipykernel==5.1.2
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
libarchive-c==2.8
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1



more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multidict==4.7.5
nbconvert==5.6.0
nbformat==4.4.0
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.16.5
numpydoc==0.9.1
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.0
opt-einsum==3.2.0
packaging==19.0
paho-mqtt==1.5.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pyreadline==2.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-crfsuite==0.9.6
python-dateutil==2.8.0
python-engineio==3.11.2
requests==2.23.0
requests-file==1.4.3
requests-oauthlib==1.3.0
rfc3986==1.3.2
rope==0.14.0
ruamel-yaml==0.15.46
s3transfer==0.3.2
sanic==19.12.2
Sanic-Cors==0.10.0.post3
sanic-jwt==1.3.2
Sanic-Plugins-Framework==0.9.2
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.4.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
tables==3.5.2
tabulate==0.8.6
tblib==1.4.0
twilio==6.26.3

